requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.46.2
Telethon==1.40.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
